% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedCC_data.R
\docType{data}
\name{cohort_test}
\alias{cohort_test}
\title{A small longitudinal cohort}
\format{A data frame with six rows and three columns:
\describe{
  \item{time}{follow-up time}
  \item{event}{status at end of follow-up (0 = no event, 1 = event)}
  \item{sex}{sex of subjects (f = female, m = male)}
}}
\usage{
cohort_test
}
\description{
A minimal cohort with follow-up time for testing
}
\references{
Hand-crafted data
}
\keyword{datasets}
