% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedCC_function.R
\name{nestedCC}
\alias{nestedCC}
\title{Generate a nested case-control design}
\usage{
nestedCC(cohort, exit, event, match, k = 1, seed)
}
\arguments{
\item{cohort}{Data frame with cohort data}

\item{exit}{Variable in data frame with event/censoring times,
specified as position (integer) or name (character)}

\item{event}{Integer variable in cohort indicating censoring (0)
or event (>0)}

\item{match}{Optional vector of variables in cohort that cases
and controls will be matched on}

\item{k}{Number of controls per case; default 1}

\item{seed}{Optional seed for random number generation}
}
\value{
A data frame with rows from the input data (possibly
        sampled repeatedly) and an extra column `matchSet`
        indicating matched sets of cases and controls.
}
\description{
Given a data frame with cohort data, this function will
return a a randomly sampled nested case-control design
}
\examples{
ex1 = nestedCC(cohort, exit = "time", event = "cvd", match = "sex", k = 1, seed = 43)
str(ex1)
head(ex1$data)
}
